Move `Try It` to `admin.applications.v1`
