+++ features/admin.extensions.v1/components/getting-started/configs/ui.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2021, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2021-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -19,7 +19,7 @@
 import { FunctionComponent, SVGProps } from "react";
 import {
     ReactComponent as TryItAppIllustration
-} from "../../../assets/illustrations/asgardeo-rafiki-illustration.svg";
+} from "../../../../themes/default/assets/images/illustrations/rafiki-illustration.svg";
 import {
     ReactComponent as OnboardUsersIllustration
 } from "../../../assets/images/illustrations/onboard-users-illustration.svg";
